pyproject.toml
[project]
name = "kurikone-cb-bot"
version = "0.1.0"
description = "Princess Connect Discord CB Bot"
requires-python = ">=3.12"
dependencies = [
    "attrs>=25.3.0",
    "dbutils>=3.1.0",
    "discord.py>=2.5.2",
    "envier>=0.6.1",
    "i18nice[yaml]>=0.15.5",
    "mariadb>=1.1.12",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-asyncio",
    "pytest-mock",
    "pytest-cov"
]

[tool.pytest.ini_options]
asyncio_mode="auto"
asyncio_default_fixture_loop_scope = "function"  # or "session", "module", etc.